+++ backend/src/main/java/co/assip/erp/depositos/cuentas_ahorro/dto/CuentaAhorroDTO.java
@@ -1,4 +1,5 @@
 package co.assip.erp.depositos.cuentas_ahorro.dto;
+import java.math.BigDecimal;
 
 import lombok.Data;
 
@@ -13,7 +14,7 @@ public class CuentaAhorroDTO {
 
     private String nombreTitular;   // ⭐ NUEVO: requerido por el Repository
 
-    private Double saldoActual;     // ⭐ NUEVO: requerido por el Repository
+    private BigDecimal saldoActual;     // ⭐ NUEVO: requerido por el Repository
 
     private String nombreForma;     // ya existía
     private String fechaApertura;   // ya existía
